change config filename from `.contentful.json` to `contentful.json`
beginner programmers and non technical people have given feedback that they can't find `.contentful.json` or `.contentful.sample.json` and we found out this is largely due to macOS' default behavior of hiding files that start with a period. This change will clarify the setup and make it easier to find and change the configuration for beginners.